All user authentication is done

+++ js/userFunctions.js
@@ -15,24 +15,63 @@ async function registerMe() {
 
 
 async function loginMe() {
+  // Get the username and password
+  const username = document.getElementById("LoginEmailUsername").value;
+  const password = document.getElementById("LoginPassword").value;
+
+  // Validate the username and password
+  if (!username) {
+    alert("Please enter a username");
+    return;
+  }
+  if (!password) {
+    alert("Please enter a password");
+    return;
+  }
+  if (username.length < 5) {
+    alert("Username must be at least 5 characters long");
+    return;
+  }
+  if (password.length < 5) {
+    alert("Password must be at least 5 characters long");
+    return;
+  }
+  if (/[^a-zA-Z0-9]/.test(username)) {
+    alert("Username must only contain alphanumeric characters");
+    return;
+  }
+  if (/[^a-zA-Z0-9]/.test(password)) {
+    alert("Password must only contain alphanumeric characters");
+    return;
+  }
+
+  // Make the fetch request if the username and password are valid
   const response = await fetch("http://127.0.0.1:8080/login", {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
     },
     body: JSON.stringify({
-      username: document.getElementById("LoginEmailUsername").value,
-      password: document.getElementById("LoginPassword").value,
+      username: username,
+      password: password,
     }),
   });
-  
+  //If successful request
   if (response.ok) {
-    const token = await response.json();
-    var username = document.getElementById("LoginEmailUsername").value;
-    localStorage.setItem("token", token.result);
-    localStorage.setItem("currentuser", username);
-    console.log(token);
-  } 
+    const token = await response.json(); //Depending on serverside, could be token or invalid
+    if (token.result != "invalid" || token.result != "Account does not exist") { //If password match
+      localStorage.setItem("token", token.result);
+      localStorage.setItem("currentuser", username);
+      localStorage.setItem("uid", token.uid);
+      console.log(token.result);
+      console.log(token.uid);
+      window.location.replace('/pages/vaultIndex.html');
+    }
+    else{ //If user doesnt exist or invalid token
+      console.log(token.result);
+    }
+
+  }
   else if (response.status == 401) {
     // Unauthorized - display an error message
     alert("Invalid username or password");
@@ -46,33 +85,5 @@ async function loginMe() {
 function logoutMe() {
   localStorage.removeItem("token");
   localStorage.removeItem("currentuser");
+  window.location.replace('/pages/vault.html');
 }
-
-
-/* function loginMe() {
-
-  var loginUser = new XMLHttpRequest();
-
-  loginUser.open("POST", "http://127.0.0.1:8080/login", true);
-  loginUser.setRequestHeader("Content-Type", "application/json");
-  loginUser.onload = function () {
-    var token = JSON.parse(loginUser.responseText); //////
-    if (token.result != false){
-      sessionStorage.setItem("token", token.result)
-      sessionStorage.setItem("currentuser", username);
-      sessionStorage.setItem("password", password);
-      console.log(token);
-    } 
-    else {
-      location.reload();
-    }
-      
-  }
-
-  var username = document.getElementById("LoginEmailUsername").value;
-  var password = document.getElementById("LoginPassword").value;
-  var payload = { username: username, password: password }
-
-  loginUser.send(JSON.stringify(payload))
-}
- */
